Reworked instruction format
Switched to array-based instruction and macro type listing (for
predefined types) to make indexing easier.

Also, removed Macro (React component) in favor of the Instruction
component armed with a flag to indicate whether or not it is a macro.

Lastly, constructed part of the Macro class, which will represent Macro
types during runtime. Finished its resolveInstructions() method —
responsible for taking instruction templates (either array or function)
and converting them into runnable (or resolvable, in the case of child
macros) instructions.